+++ horde/classes/base/processing_generation.py
@@ -12,6 +12,10 @@ uuid_column_type = lambda: UUID(as_uuid=True) if not SQLITE_MODE else db.String(
 class ProcessingGeneration(db.Model):
     """For storing processing generations in the DB"""
     __tablename__ = "processing_gens"
+    __mapper_args__ = {
+        "polymorphic_identity": "template",
+        "polymorphic_on": "procgen_type",
+    }    
     id = db.Column(uuid_column_type(), primary_key=True, default=get_db_uuid)
     generation = db.Column(db.Text)
 
@@ -86,7 +90,6 @@ class ProcessingGeneration(db.Model):
             logger.info(f"Fake{cancel_txt} Generation {self.id} worth {self.kudos} kudos, delivered by worker: {self.worker.name} for wp {self.wp.id}")
         else:
             self.worker.record_contribution(raw_things = self.wp.things, kudos = kudos, things_per_sec = things_per_sec)
-            
             self.wp.record_usage(raw_things = self.wp.things, kudos = self.adjust_user_kudos(kudos))
             logger.info(f"New{cancel_txt} Generation {self.id} worth {kudos} kudos, delivered by worker: {self.worker.name} for wp {self.wp.id}")