+++ horde/classes/kobold/processing_generation.py
@@ -1,0 +1,45 @@
+from horde.logger import logger
+from horde.classes.base.processing_generation import ProcessingGeneration
+from horde.classes.kobold.genstats import record_text_statistic
+from horde.flask import db
+from horde.model_reference import model_reference
+
+
+class ProcessingGenerationExtended(ProcessingGeneration):
+    censored = db.Column(db.Boolean, default=False, nullable=False)
+
+    def get_details(self):
+        '''Returns a dictionary with details about this processing generation'''
+        ret_dict = {
+            "text": self.generation,
+            "seed": self.seed,
+            "worker_id": self.worker.id,
+            "worker_name": self.worker.name,
+            "model": self.model,
+            "id": self.id,
+        }
+        return ret_dict
+
+    def get_gen_kudos(self):
+        # We have pre-calculated them as they don't change per worker        
+        return round(self.wp.tokens * model_reference.get_text_model_multiplier(self.model) / 21, 2)
+
+
+    def log_aborted_generation(self):
+        record_text_statistic(self)
+        logger.info(
+            f"Aborted Stale Generation {self.id} "
+            f"({self.wp.width}x{self.wp.height}x{self.wp.params['steps']}@{self.wp.params['sampler_name']})"
+            f" from by worker: {self.worker.name} ({self.worker.id})"
+        )
+
+
+    def set_generation(self, generation, things_per_sec, **kwargs):
+        # We don't check the state in the super() function as image gen sets it early here
+        # as well, so it can abort before doing R2 operations
+        state = kwargs.get("state", 'ok')
+        if state == "faulted":
+            self.abort()
+        kudos = super().set_generation(generation, things_per_sec, **kwargs)
+        record_text_statistic(self)
+        return(kudos)