+++ horde/utils.py
@@ -83,3 +83,4 @@ def get_interrogation_form_expiry_date():
 def get_random_seed():
     '''Generated a random seed, using a random number unique per node'''
     return random.randint(0, 2**32 - 1)
+